test(dpp): move ranked-index update tests to validate_update v1
The ranked aggregate index grammar (#4266) only exists at protocol v14,
where validate_update now dispatches to v1, so the rankedAverageable
update tests added to the v0 module assert the v1 rejection shape
("changed index 'byRestaurant'") and belong in the v1 module. The
ranked flags ride the same name-keyed index-definition comparison as
every other index property.
