+++ packages/rs-dpp/src/data_contract/document_type/methods/validate_update/v0/mod.rs
@@ -147,108 +147,4 @@ mod tests {
             .validate_update(new_late_name.as_ref(), platform_version)
             .expect_err("late-name addition should error in schema compatibility");
     }
-
-    /// The ranking axes are index-level, so `validate_config` — which covers
-    /// the *doctype*-level count / sum flags just above — is deliberately not
-    /// where they are enforced. They ride the index-structure comparison
-    /// (`IndexLevel::validate_update`) that `validate_update_v0` runs right
-    /// after the config check. These tests exercise the PUBLIC dispatcher so
-    /// that routing is pinned, not just the helper in isolation.
-    mod validate_update_ranked_indices {
-        use super::*;
-        use std::collections::BTreeMap;
-
-        /// `review` doctype, one averageable index over `restaurantId`, with
-        /// `rankedAverageable` set to the supplied value.
-        fn document_type_with_ranked_index(
-            ranked_averageable: bool,
-            platform_version: &PlatformVersion,
-        ) -> DocumentType {
-            let schema = platform_value!({
-                "type": "object",
-                "properties": {
-                    // 32 rather than the generic 63-character index limit:
-                    // an index declaring a ranking axis bounds its group key
-                    // more tightly (59 characters on the Avg axis), and both
-                    // halves of these tests have to build the same doctype
-                    // shape with only `rankedAverageable` differing.
-                    "restaurantId": {
-                        "type": "string",
-                        "maxLength": 32,
-                        "position": 0,
-                    },
-                    "grade": {
-                        "type": "integer",
-                        "minimum": 0,
-                        "maximum": 100,
-                        "position": 1,
-                    },
-                },
-                "required": ["restaurantId", "grade"],
-                "additionalProperties": false,
-                "indices": [{
-                    "name": "byRestaurant",
-                    "properties": [{ "restaurantId": "asc" }],
-                    "averageable": "grade",
-                    "rangeAverageable": true,
-                    "rankedAverageable": ranked_averageable,
-                }],
-            });
-
-            let config = DataContractConfig::default_for_version(platform_version)
-                .expect("should create a default config");
-
-            DocumentType::try_from_schema(
-                Identifier::random(),
-                1,
-                config.version(),
-                "review",
-                schema,
-                None,
-                &BTreeMap::new(),
-                &config,
-                true,
-                &mut Vec::new(),
-                platform_version,
-            )
-            .expect("failed to create document type")
-        }
-
-        #[test]
-        fn should_return_invalid_result_when_ranked_averageable_is_changed() {
-            let platform_version = PlatformVersion::latest();
-            let old = document_type_with_ranked_index(false, platform_version);
-            let new = document_type_with_ranked_index(true, platform_version);
-
-            let result = old
-                .as_ref()
-                .validate_update(new.as_ref(), platform_version)
-                .expect("validate_update should not error");
-
-            assert_matches!(
-                result.errors.as_slice(),
-                [ConsensusError::BasicError(
-                    crate::consensus::basic::BasicError::DataContractInvalidIndexDefinitionUpdateError(e)
-                )] if e.index_path() == "restaurantId -> (ranked_averageable: false -> true)"
-            );
-        }
-
-        #[test]
-        fn should_pass_when_ranked_averageable_is_unchanged() {
-            let platform_version = PlatformVersion::latest();
-            let old = document_type_with_ranked_index(true, platform_version);
-            let new = document_type_with_ranked_index(true, platform_version);
-
-            let result = old
-                .as_ref()
-                .validate_update(new.as_ref(), platform_version)
-                .expect("validate_update should not error");
-
-            assert!(
-                result.is_valid(),
-                "an unchanged ranked index must not be rejected, got {:?}",
-                result.errors
-            );
-        }
-    }
 }
